refactores tests: complement, compose, nth, one
