+++ tests/test_itertools_cfuncs.py
@@ -14,306 +14,6 @@ from helper_leak import memory_leak
 from helper_pytest_monkeypatch import pytest_raises
 
 
-def test_complement():
-    complement = iteration_utilities.complement
-
-    assert not complement(lambda x: x is True)(True)
-    assert complement(lambda x: x is True)(False)
-
-    assert complement(lambda x: x is False)(True)
-    assert not complement(lambda x: x is False)(False)
-
-    assert not complement(iteration_utilities.is_None)(None)
-    assert complement(iteration_utilities.is_None)(False)
-    assert complement(iteration_utilities.is_None)(True)
-
-
-def test_complement_memoryleak():
-    complement = iteration_utilities.complement
-
-    def test():
-        complement(lambda x: x is True)(True)
-    assert not memory_leak(test)
-
-    def test():
-        complement(lambda x: x is True)(False)
-    assert not memory_leak(test)
-
-    def test():
-        complement(lambda x: x is False)(True)
-    assert not memory_leak(test)
-
-    def test():
-        complement(lambda x: x is False)(False)
-    assert not memory_leak(test)
-
-    def test():
-        complement(iteration_utilities.is_None)(None)
-    assert not memory_leak(test)
-
-    def test():
-        complement(iteration_utilities.is_None)(False)
-    assert not memory_leak(test)
-
-    def test():
-        complement(iteration_utilities.is_None)(True)
-    assert not memory_leak(test)
-
-
-def test_compose():
-    compose = iteration_utilities.compose
-
-    double_increment = compose(lambda x: x*2, lambda x: x+1)
-    assert double_increment(10) == 21
-    assert double_increment(2.) == 5
-
-    with pytest.raises(TypeError):  # at least one func must be present
-        compose()
-
-    with pytest.raises(TypeError):  # kwarg not accepted
-        compose(lambda x: x+1, invalidkwarg=lambda x: x*2)
-
-    with pytest.raises(TypeError):  # func fails
-        compose(lambda x: x+1)('a')
-
-    with pytest.raises(TypeError):  # second func fails
-        compose(lambda x: x*2, lambda x: x+1)('a')
-
-
-def test_compose_memoryleak():
-    compose = iteration_utilities.compose
-
-    class Test(object):
-        def __init__(self, value):
-            self.value = value
-
-        def __add__(self, other):
-            return self.__class__(self.value + other)
-
-        def __mul__(self, other):
-            return self.__class__(self.value + other)
-
-    def test():
-        compose(lambda x: x*2, lambda x: x+1)(Test(10))
-        compose(lambda x: x*2, lambda x: x+1)(Test(2))
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(TypeError):  # at least one func must be present
-            compose()
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(TypeError):  # kwarg not accepted
-            compose(lambda x: x+1, invalidkwarg=lambda x: x*2)
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(TypeError):  # func fails
-            compose(lambda x: x+1)(Test('a'))
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(TypeError):  # second func fails
-            compose(lambda x: x*2, lambda x: x+1)(Test('a'))
-    assert not memory_leak(test)
-
-
-def test_one():
-    one = iteration_utilities.one
-
-    assert one([0]) == 0
-    assert one('a') == 'a'
-    assert one({'o': 10}) == 'o'
-
-    with pytest.raises(TypeError):
-        one(10)
-
-    with pytest.raises(ValueError):  # empty iterable
-        one([])
-
-    with pytest.raises(ValueError):  # more than 1 element
-        one([1, 2])
-
-
-def test_one_memoryleak():
-    one = iteration_utilities.one
-
-    class Test(object):
-        def __init__(self, value):
-            self.value = value
-
-    def test():
-        one([Test(0)])
-    assert not memory_leak(test)
-
-    def test():
-        one('a')
-    assert not memory_leak(test)
-
-    def test():
-        one({Test(0): 10})
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(TypeError):
-            one(Test(0))
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(ValueError):  # empty iterable
-            one([])
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(ValueError):  # more than 1 element
-            one([1, 2])
-    assert not memory_leak(test)
-
-
-def test_nth():
-    nth = iteration_utilities.nth
-
-    assert nth([1, 2, 3], 1) == 2
-    assert nth(range(10), 2) == 2
-
-    # With pred
-    assert nth([0, 1, 2], 1, pred=bool) == 2
-    assert nth([0, 1, 2], 1, pred=None) == 2
-    assert nth([0]*100 + [1], 0, pred=bool) == 1
-    assert nth([[1], [1, 2], [1, 2]], 1, pred=lambda x: len(x) > 1) == [1, 2]
-
-    # pred with truthy/retpred
-    assert nth([0, 1, 2, 3, 0], 1, pred=bool, truthy=False) == 0
-    assert nth([0, 1, 2, 3, 0], 1, pred=bool, truthy=False,
-                retpred=True) == False
-    assert nth([0, 1, 2, 3, 0], 1, pred=lambda x: x**2, truthy=False) == 0
-    assert nth([0, 1, 2, 3, 0], 1,
-               pred=lambda x: x**2, truthy=False, retpred=True) == 0
-    assert nth([0, 1, 2, 3], 2, pred=bool) == 3
-    assert nth([0, 1, 2, 3], 2, pred=bool, retpred=True) == True
-    assert nth([0, 2, 3, 4], 2, pred=lambda x: x**2) == 4
-    assert nth([0, 2, 3, 4], 2, pred=lambda x: x**2, retpred=True) == 16
-
-    # With default
-    assert nth([], 2, default=None) is None
-    assert nth([0, 0, 0], 1, default=None, pred=bool) is None
-
-    # failures
-    with pytest.raises(TypeError):
-        nth(100, 10)
-
-    with pytest.raises(IndexError):
-        nth([], 10)
-
-    with pytest.raises(IndexError):
-        nth([0], 1, pred=bool)
-
-    with pytest.raises(TypeError):
-        nth(['a', 'b'], 1, pred=abs)
-
-
-def test_nth_memoryleak():
-    nth = iteration_utilities.nth
-
-    class Test(object):
-        def __init__(self, value):
-            self.value = value
-
-        def __bool__(self):
-            return bool(self.value)
-
-        def __nonzero__(self):
-            return bool(self.value)
-
-        def __pow__(self, other):
-            return self.__class__(self.value ** other.value)
-
-    def test():
-        nth([Test(1), Test(2), Test(3)], 1)
-    assert not memory_leak(test)
-
-    # With pred
-    def test():
-        nth([Test(0), Test(1), Test(2)], 1, pred=bool)
-    assert not memory_leak(test)
-
-    def test():
-        nth([Test(0), Test(1), Test(2)], 1, pred=None)
-    assert not memory_leak(test)
-
-    def test():
-        nth([Test(0)]*100 + [Test(1)]*2, 1, pred=bool)
-    assert not memory_leak(test)
-
-    def test():
-        nth([[Test(0)], [Test(1), Test(2)]]*2, 1, pred=lambda x: len(x) > 1)
-    assert not memory_leak(test)
-
-    # pred with truthy/retpred
-    def test():
-        nth([Test(0), Test(2), Test(3), Test(0)], 1,
-            pred=bool, truthy=False)
-    assert not memory_leak(test)
-
-    def test():
-        nth([Test(0), Test(2), Test(3), Test(0)], 1,
-            pred=bool, truthy=False, retpred=True)
-    assert not memory_leak(test)
-
-    def test():
-        nth([Test(0), Test(2), Test(3), Test(0)], 1,
-            pred=lambda x: x**Test(2), truthy=False)
-    assert not memory_leak(test)
-
-    def test():
-        nth([Test(0), Test(2), Test(3), Test(0)], 0,
-            pred=lambda x: x**Test(2), truthy=False, retpred=True)
-    assert not memory_leak(test)
-
-    def test():
-        nth([Test(0), Test(2), Test(3)], 1, pred=bool)
-    assert not memory_leak(test)
-
-    def test():
-        nth([Test(0), Test(2), Test(3)], 1, pred=bool, retpred=True)
-    assert not memory_leak(test)
-
-    def test():
-        nth([Test(0), Test(2), Test(3)], 1, pred=lambda x: x**Test(2))
-    assert not memory_leak(test)
-
-    def test():
-        nth([Test(0), Test(2), Test(3)], 1,
-            pred=lambda x: x**Test(2), retpred=True)
-    assert not memory_leak(test)
-
-    # With default
-    def test():
-        nth([], 2, default=None) is None
-    assert not memory_leak(test)
-
-    def test():
-        nth([Test(0), Test(0), Test(0)], 1, default=None, pred=bool) is None
-    assert not memory_leak(test)
-
-    # failures
-    def test():
-        with pytest_raises(IndexError):
-            nth([], 10)
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(TypeError):
-            nth(Test(100), 10)
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(TypeError):
-            nth([Test('a'), Test('b')], 1, pred=lambda x: abs(x.value))
-    assert not memory_leak(test)
-
-
 def test_quantify():
     quantify = iteration_utilities.quantify
